Move non `@Composable` methods from `ui/Common.kt` to `model/Utils.kt`
