+++ app/src/main/java/com/github/se/signify/ui/screens/home/HomeScreen.kt
@@ -53,13 +53,13 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.github.se.signify.R
 import com.github.se.signify.model.exercise.ExerciseLevel
+import com.github.se.signify.model.getIconResId
+import com.github.se.signify.model.getImageResId
+import com.github.se.signify.model.getLetterIconResId
+import com.github.se.signify.model.getTipResId
 import com.github.se.signify.ui.MainScreenScaffold
 import com.github.se.signify.ui.UtilButton
 import com.github.se.signify.ui.UtilTextButton
-import com.github.se.signify.ui.getIconResId
-import com.github.se.signify.ui.getImageResId
-import com.github.se.signify.ui.getLetterIconResId
-import com.github.se.signify.ui.getTipResId
 import com.github.se.signify.ui.navigation.NavigationActions
 import com.github.se.signify.ui.navigation.Screen
 import kotlinx.coroutines.CoroutineScope
